requirements.txt
# Core Framework
dash==2.14.2
dash-bootstrap-components==1.5.0
plotly==5.18.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Geospatial
geopandas==0.14.1
shapely==2.0.2

# Caching & Performance
redis==5.0.1
flask-caching==2.1.0
diskcache==5.6.3

# Utils
python-dotenv==1.0.0
pyyaml==6.0.1

# Production Server
gunicorn==21.2.0
waitress==2.1.2

# Data Validation
pydantic==2.5.3

# Scientific Computing
scikit-learn==1.3.2
statsmodels==0.14.1

# Visualization Enhancements
kaleido==0.2.1
colorlover==0.3.0
